+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Lowercases the 'r' of '(required)' in labels and placeholders where necessary in order to align with sentence case guidelines.
-- Adds 'There was a problem. ' prefix on invalid text in the SearchBar component.
+- Updates invalid text in the SearchBar component to include "There was a problem. " prefix.
 
 ### Fixes